Refactoring + Clear Task information endpoints
