+++ bazelwebapi/BUILD.bazel
@@ -7,14 +7,16 @@ go_library(
     name = "bazelwebapi_lib",
     srcs = [
         "config.go",
-        "executebazel.go",
         "git.go",
+        "handlers.go",
         "main.go",
         "middleware.go",
         "prio.go",
+        "queue.go",
         "routes.go",
         "server.go",
         "state.go",
+        "task.go",
     ],
     importpath = "github.com/Weiss91/web-cd-service/bazelwebapi",
     visibility = ["//visibility:private"],
